+++ src/main/kotlin/net/fallingangel/jimmerdto/highlighting/DTOAnnotator.kt
@@ -402,6 +402,7 @@ class DTOAnnotator : Annotator {
             val type = o.type.value
             val clazz = o.type.clazz
 
+            // 类型解析
             if (clazz == null && type !in DTOLanguage.preludes) {
                 o.type.error(
                     "Unresolved reference: $type",
@@ -410,6 +411,7 @@ class DTOAnnotator : Annotator {
                 return
             }
 
+            // 泛型校验
             val exceptedTypeParamNumber = GenericType[type]?.generics?.size ?: clazz?.typeParameters?.size ?: 0
             if ((o.arguments?.values?.size ?: 0) != exceptedTypeParamNumber) {
                 o.type.error("Generic parameter mismatch")
@@ -430,6 +432,12 @@ class DTOAnnotator : Annotator {
                     RenameElement(o.type, Project::createUserPropType),
                 )
                 return
+            }
+
+            // 默认值校验
+            val dto = o.parentOfType<DTODto>() ?: return
+            if (o.questionMark == null && dto notModifiedBy Modifier.Specification && type !in DTOLanguage.preludes) {
+                o.type.error("Type is not null and its default value cannot be determined")
             }
         }